Transition-Validate (#14)
* plotting
Added:
- transition_model.pdf
- config_plot.py
Modified:
- manhattan-cleaned.graphml
- manhattan-raw.graphml
- manhattan-sparse-782-nodes-enriched.graphml
- manhattan-sparse-782-nodes.graphml
- SAVE-MAP.ipynb
- transition_model.ipynb
- scratch.ipynb
- __init__.py
- config.py
- requests.py
- vehicles.py
- ENV.py
- graph_mixin.py
- interface.py
- obs_space_mixin.py
- rendering_mixin.py
- transition_mixin.py
- __init__.py
- euclidean_L2_embed.py
- visualization.py

* delete
Added:
- test_tick_labels.png
- test_tick_labels.py
Modified:
- transition_model.ipynb
- visualization.py

* plot-good
Modified:
- transition_model.ipynb
- scratch.ipynb
- __init__.py
- active_rides.py
- config.py
- config_plot.py
- enriched_df_base.py
- metrics.py
- requests.py
- vehicles.py
- ENV.py
- cost_reward.py
- graph_mixin.py
- interface.py
- obs_space_mixin.py
- rendering_mixin.py
- transition_mixin.py
- __init__.py
- euclidean_L2_embed.py
- inspect.py
- load_graph_safe.py
- post_process_enrichment.py
- pre_processing.py
- sparsify.py
- traverse_graph.py
- visualization.py
Deleted:
- test_tick_labels.py

* plot-good
Modified:
- scratch.ipynb

* isort
Modified:
- scratch.ipynb
- config_plot.py
- requests.py
- ENV.py
- obs_space_mixin.py
- rendering_mixin.py
- visualization.py

* transition-model-working
Modified:
- transition_model.ipynb
- scratch.ipynb
- config.py
- customer.py
- cost_reward.py
- df_utils.py
- action_mixin.py
- obs_space_mixin.py
- rendering_mixin.py
- transition_mixin.py
- traverse_graph.py
- generate_obs_state.py

+++ waymo_agent/graph_env/mixin/action_mixin.py
@@ -9,6 +9,7 @@ from gymnasium import spaces
 from waymo_agent.data_classes import ActionDict, validate_keys
 
 from .interface import GymEnvInterface
+import numpy as np
 
 
 class ActionMixin(GymEnvInterface):
@@ -29,7 +30,9 @@ class ActionMixin(GymEnvInterface):
         """
         action_space = spaces.Dict(
             {
-                "prices": spaces.Box(low=0.0, high=self.config.max_price, shape=(self.config.max_pending_requests,)),
+                "prices": spaces.Box(
+                    low=0.0, high=self.config.max_price, shape=(self.config.max_pending_requests,), dtype=np.float64
+                ),
                 "reposition": spaces.Box(low=-1.0, high=1.0, shape=(self.num_vehicles, 2)),
                 "dispatch": spaces.MultiDiscrete(
                     [self.num_vehicles + 1] * self.config.max_pending_requests,
